+++ fltk/src/menu.rs
@@ -129,36 +129,6 @@ impl MenuItem {
         }
     }
 
-    /// Initializes a new menu item, with the added menu flag
-    pub fn new_ext(choices: &'static [Option<(&str, Shortcut, MenuFlag)>]) -> MenuItem {
-        unsafe {
-            let sz = choices.len();
-            let mut labels: Vec<*mut raw::c_char> = vec![];
-            let mut shortcuts: Vec<i32> = vec![];
-            let mut flags: Vec<i32> = vec![];
-            for &choice in choices {
-                if let Some(choice) = choice {
-                    let c = CString::safe_new(choice.0);
-                    labels.push(c.into_raw());
-                    shortcuts.push(choice.1.bits() as i32);
-                    flags.push(choice.2 as i32);
-                } else {
-                    labels.push(std::ptr::null_mut());
-                    shortcuts.push(0 as i32);
-                    flags.push(0 as i32);
-                }
-            }
-            let item_ptr = Fl_Menu_Item_new_ext(
-                labels.as_ptr() as *mut *mut raw::c_char,
-                shortcuts.as_mut_ptr(),
-                flags.as_mut_ptr(),
-                sz as i32,
-            );
-            assert!(!item_ptr.is_null());
-            MenuItem { inner: item_ptr }
-        }
-    }
-
     /// Creates a popup menu at the specified coordinates and returns its choice
     pub fn popup(&self, x: i32, y: i32) -> Option<MenuItem> {
         assert!(!self.was_deleted());