+++ fltk/src/app/visual.rs
@@ -95,6 +95,34 @@ pub fn swap_frame_type(new_frame: FrameType) {
     }
 }
 
+/// Get the shadwo width for frames types with shadows
+pub fn frame_shadow_width() -> i32 {
+    unsafe {
+        fl::Fl_box_shadow_width()
+    }
+}
+
+/// Set the shadwo width for frames types with shadows
+pub fn set_frame_shadow_width(width: i32) {
+    unsafe {
+        fl::Fl_set_box_shadow_width(width)
+    }
+}
+
+/// Get the max border radius for frame types
+pub fn frame_border_radius_max() -> i32 {
+    unsafe {
+        fl::Fl_box_border_radius_max()
+    }
+}
+
+/// Set the max border radius for frame types
+pub fn set_frame_border_radius_max(radius: i32) {
+    unsafe {
+        fl::Fl_set_box_border_radius_max(radius)
+    }
+}
+
 /// Makes FLTK use its own colormap. This may make FLTK display better
 pub fn own_colormap() {
     unsafe { fl::Fl_own_colormap() }