+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## [1.1.3] - Unreleased
 - Add WidgetExt::with_type().
-- Add MenuExt::menu and set_menu().
+- Add MenuExt::menu.
 
 ## [1.1.2] - 2021-07-19
 - Fix app::event_clicks() and add overload for event_clicks_num() for number of clicks.